+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Bugfix: Char boundary string splitting error
 - Allow dynamic prompt made from iterator over `&str`
+- Take slice instead of owning buffer for line and history
 
 ## [0.4.0 - 2024-08-27]